Move TopBar composable into "designsystem" package.
